added main menu and base
Added a base at the end of the track so the enemies aren't disappearing for no reason. in the future it will also be used to house drops before being picked up.

Added simple main menu with 2 buttons. Start and Exit. Start launches the LevelScene and Exit exits the game.